- added PalauEntry Protocol and PalauCustomDefaultable protocol
PalauEntry Protocol is used as base for PalauDefaultsEntry and
PalauDefaultsArrayEntry. Most of the methods can be defined in an
extension of PalauEntry.

PalauCustomDefaultable is used to provide easy support for e.g. Struct
mapping either for PalauDefaultsEntry or PalauDefaultsArrayEntry.
The implementation of PalauCustomDefaultable needs to provide a
toIntermediate and fromIntermediate func.